Add sidebar
We don't have access to the docsearch account that this project uses, so for now, I'm disabling the search box. It'll be re-enabled when the work in #195 is completed.

Note that this PR adds the `sidebar.html` partial from `hugo-cloudflare-docs`, meaning that we'll own our sidebar config - this will give us more control over how pages are formatted there - woo!