Install pinned Hugo version in CI
This fixes a current issue in the deployed version of our docs where the newest version of hugo is unable to build both of our `index` files in `content/`. Building without the placeholder `content/_index.html` causes our sidebar to break. Since we'll eventually be moving off of this codebase, I'm OK to just pin a known-good version of Hugo here and build with that (the last patch-level version for the 0.59.x series - 0.59.1)

Note that this reverts some work from #489 - we now know that we want to be on a newer version of Hugo then prior to that commit, but not so new that we move to 0.60.0 and have to fix a bunch of things in the codebase.