Fixes Hugo-style `ref` links in `/archive` docs
A number of pages in the /archive section have broken relative links of
the form `[Resource Limits]({{< ref
"archive/writing-workers/resource-limits.md" >}})`. That appears to be a
Hugo thing: https://gohugo.io/content-management/cross-references/ .
This change replaces all such links.

Also, updated the formatting on the code block on the /archive/reference
page while I was there, because it was broken.